+++ src/main/java/com/mobble/mobbleserver/refactor/meeting/dto/request/MeetingUpdateRequestDto.java
@@ -1,7 +1,7 @@
 package com.mobble.mobbleserver.refactor.meeting.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.mobble.mobbleserver.refactor.meeting.entity.MeetingType;
+import com.mobble.mobbleserver.infrastructure.web.meeting.MeetingType;
 
 import java.time.LocalDateTime;
