+++ src/main/java/com/mobble/mobbleserver/refactor/meeting/dto/response/MeetingResponseDto.java
@@ -1,8 +1,8 @@
 package com.mobble.mobbleserver.refactor.meeting.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.mobble.mobbleserver.refactor.meeting.entity.Meeting;
-import com.mobble.mobbleserver.refactor.meeting.entity.MeetingType;
+import com.mobble.mobbleserver.infrastructure.web.meeting.Meeting;
+import com.mobble.mobbleserver.infrastructure.web.meeting.MeetingType;
 
 import java.time.LocalDateTime;
